feat: スキルシステム、レベルシステム、ジェネレーターシステム、Oraxen用アイテム管理、コンフィグDI、命名整備、

+++ src/main/kotlin/com/github/ryuzu/ryuzutechnicalmagiccore/core/model/configuration/effect/particle/ConfiguredParticle.kt
@@ -1,11 +1,11 @@
 package com.github.ryuzu.ryuzutechnicalmagiccore.core.model.configuration.effect.particle
 
-import com.github.ryuzu.ryuzutechnicalmagiccore.core.model.configuration.ConfiguredDoubleVector
+import com.github.ryuzu.ryuzutechnicalmagiccore.core.model.configuration.base.ConfiguredDoubleVector
 
 data class ConfiguredParticle(
     override val id: String,
     override val count: Int,
-    override val extra: Float,
+    override val extra: Double,
     override val offset: ConfiguredDoubleVector,
-    override val delay: Int
+    override val delay: Long
 ) : IConfiguredParticle
